+++ R/criteriobrasil-package.R
@@ -1,5 +1,8 @@
 #' criteriobrasil
 #'
+#' @description
+#' `r lifecycle::badge("experimental")`
+#'
 #' Tidy tables from the Brazilian Economic Classification Criterion (CCEB),
 #' including scoring rules, class cutoffs, income estimates, and geographic
 #' distributions.
@@ -8,5 +11,6 @@
 "_PACKAGE"
 
 ## usethis namespace: start
+#' @importFrom lifecycle deprecated
 ## usethis namespace: end
 NULL